+++ src/modules/serviceManager/index.ts
@@ -32,7 +32,7 @@ function maskConfig(config) {
         break;
       case 'interactiveAuth':
         if (Array.isArray(configValue)) {
-          copy[key] = configValue.map(phrase => MASK);
+          copy[key] = configValue.map(() => MASK);
         } else {
           copy[key] = configValue;
         }